separate stale file handle tests to separate file for streaming writes.
