+++ internal/bufferedwrites/upload_handler.go
@@ -114,25 +114,21 @@ func (uh *UploadHandler) createObjectWriter() (err error) {
 	return
 }
 
-func (uh *UploadHandler) getUploadError() error {
+func (uh *UploadHandler) UploadError() error {
 	return uh.uploadError
-}
-
-func (uh *UploadHandler) setUploadError(err error) {
-	uh.uploadError = err
 }
 
 // uploader is the single-threaded goroutine that uploads blocks.
 func (uh *UploadHandler) uploader() {
 	for currBlock := range uh.uploadCh {
-		if uh.getUploadError() != nil {
+		if uh.UploadError() != nil {
 			uh.wg.Done()
 			continue
 		}
 		_, err := io.Copy(uh.writer, currBlock.Reader())
 		if err != nil {
 			logger.Errorf("buffered write upload failed for object %s: error in io.Copy: %v", uh.objectName, err)
-			uh.setUploadError(gcs.GetGCSError(err))
+			uh.uploadError = gcs.GetGCSError(err)
 		}
 		// Put back the uploaded block on the freeBlocksChannel for re-use.
 		uh.freeBlocksCh <- currBlock
@@ -157,7 +153,7 @@ func (uh *UploadHandler) Finalize() (*gcs.MinObject, error) {
 	obj, err := uh.bucket.FinalizeUpload(context.Background(), uh.writer)
 	if err != nil {
 		// FinalizeUpload already returns GCSerror so no need to convert again.
-		uh.setUploadError(err)
+		uh.uploadError = err
 		logger.Errorf("FinalizeUpload failed for object %s: %v", uh.objectName, err)
 		return nil, err
 	}
